move my account static experience data to labs
