+++ libs/template/labs/src/feature.ts
@@ -4,6 +4,7 @@ import { bazaarVoiceComponentMapping } from './bazaarvoice';
 import { cloudinaryImageConverter } from './cloudinary';
 import * as components from './components';
 import { i18nLabsProviders, labsI18nFeature } from './i18n';
+import { myAccountFeature } from './my-account';
 export * from './components';
 export { labsI18nFeature } from './i18n';
 
@@ -17,6 +18,7 @@ export const labsComponents = Object.values(components);
  */
 export const labsFeatures: AppFeature[] = [
   labsI18nFeature,
+  myAccountFeature,
   {
     components: labsComponents,
     providers: [
