+++ libs/template/labs/src/my-account/my-account-navigation.ref.ts
@@ -1,6 +1,6 @@
 import { ExperienceComponent } from '@spryker-oryx/experience';
 import { IconTypes } from '@spryker-oryx/ui/icon';
-import { Size, i18n } from '@spryker-oryx/utilities';
+import { i18n } from '@spryker-oryx/utilities';
 
 export const myAccountNavigation: ExperienceComponent = {
   type: 'oryx-composition',
@@ -67,50 +67,4 @@ export const myAccountNavigation: ExperienceComponent = {
       },
     ],
   },
-};
-
-export const myAccountPage: ExperienceComponent = {
-  id: 'my-account',
-  type: 'Page',
-  meta: {
-    title: 'My account',
-    route: '/my-account',
-  },
-  components: [
-    { ref: 'header' },
-    {
-      type: 'oryx-composition',
-      options: {
-        rules: [
-          {
-            layout: { type: 'split', columnWidthType: 'aside' },
-            padding: '30px 0 0',
-          },
-        ],
-      },
-      components: [
-        {
-          type: 'oryx-site-breadcrumb',
-          options: {
-            rules: [
-              {
-                colSpan: 2,
-              },
-              { query: { breakpoint: Size.Sm }, hide: true },
-            ],
-          },
-        },
-        { ref: 'myAccountNavigation' },
-        {
-          type: 'oryx-content-text',
-          content: {
-            data: {
-              text: `<h1>My account</h1><p>content...</p>`,
-            },
-          },
-        },
-      ],
-    },
-    { ref: 'footer' },
-  ],
 };
