+++ libs/template/presets/storefront/experience/experience-data.ts
@@ -13,8 +13,6 @@ import {
   editAddressPage,
   homePage,
   loginPage,
-  myAccountNavigation,
-  myAccountPage,
   orderConfirmationPage,
   productPage,
   registrationPage,
@@ -35,8 +33,6 @@ export const StaticExperienceFeature: AppFeature = {
       orderConfirmationPage,
       productPage,
       searchPage,
-      myAccountPage,
-      myAccountNavigation,
       addressBookPage,
       createAddressPage,
       editAddressPage,
