+++ libs/template/presets/storefront/experience/pages/my-account/index.ts
@@ -1,4 +1,3 @@
 export * from './addresses.page';
 export * from './create-address.page';
 export * from './edit-address.page';
-export * from './my-account.page';
